rename(route-folder): main페이지 components 폴더 컨벤션 반영
