+++ src/app/(home)/_components/MainSearchHeader/MainSearchHeader.tsx
@@ -7,7 +7,7 @@ import { cn } from "@/utils";
 import type { KeyboardEvent as ReactKeyboardEvent, RefObject } from "react";
 import { Suspense, useEffect, useRef, useState } from "react";
 import SearchFocusDropdown from "../SearchFocusDropdown/SearchFocusDropdown";
-import { focusSearchDropdownBoundary } from "../SearchFocusDropdown/_internal/searchDropdownListKeyboard";
+import { focusSearchDropdownBoundary } from "../../_utils/searchDropdownListKeyboard";
 import MainSearchLayout from "../MainSearchLayout/MainSearchLayout";
 import { DEFAULT_ADDRESS } from "@/constants";
 import { useGeolocationPermissionGranted } from "@/hooks";
